#6 refactor : APiState response -> model로 패키지 이동
