+++ Aos/app/src/main/java/com/avengers/nibobnebob/data/repository/LoginRepository.kt
@@ -1,6 +1,6 @@
 package com.avengers.nibobnebob.data.repository
 
-import com.avengers.nibobnebob.data.model.response.ApiState
+import com.avengers.nibobnebob.data.model.ApiState
 import com.avengers.nibobnebob.data.model.response.NaverLoginResponse
 import kotlinx.coroutines.flow.Flow
