+++ Aos/app/src/main/java/com/avengers/nibobnebob/data/repository/LoginRepositoryImpl.kt
@@ -1,7 +1,7 @@
 package com.avengers.nibobnebob.data.repository
 
 import com.avengers.nibobnebob.app.DataStoreManager
-import com.avengers.nibobnebob.data.model.response.ApiState
+import com.avengers.nibobnebob.data.model.ApiState
 import com.avengers.nibobnebob.data.model.response.NaverLoginResponse
 import com.avengers.nibobnebob.data.remote.NibobNebobApi
 import kotlinx.coroutines.flow.Flow
