+++ Aos/app/src/main/java/com/avengers/nibobnebob/presentation/ui/intro/login/LoginViewModel.kt
@@ -4,7 +4,7 @@ import android.util.Log
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import com.avengers.nibobnebob.app.DataStoreManager
-import com.avengers.nibobnebob.data.model.response.ApiState
+import com.avengers.nibobnebob.data.model.ApiState
 import com.avengers.nibobnebob.data.repository.LoginRepository
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.flow.MutableSharedFlow
